Counties are re-ordered alphabetically when searching. DRY'd up county filter code
