+++ app/views/serviceProviderFilter.xml
@@ -32,10 +32,7 @@
 					<Label>COUNTIES</Label>
 					<ImageView class="arrow" id="leftArrow" left="10dp" image="/global/back-arrow.png" onClick="gotoPrevious" />
 				</View>
-				<SearchBar platform="android"  id="searchBar" height="50dp" backgroundColor="#D2B48C"/>
-				<ScrollView scrollType="vertical" backgroundColor="#fff">
-					<View id="countiesPicker"/>
-				</ScrollView>
+				<Require id="countySelector" src="countySelectorView" />
 			</View>
 		</ScrollableView>
 	</Window>
